Added BI tests... and logging out... it can be improved.

+++ bsbm.sh
@@ -17,10 +17,11 @@
 ##
 
 BSBM_ROOT_PATH=/tmp/bsbm
-BSBM_SCALE_FACTOR=10000
+BSBM_SCALE_FACTOR=14092
 BSBM_NUM_QUERY_MIXES=128
 BSBM_NUM_QUERY_WARM_UP=32
-BSBM_CONCURRENT_CLIENTS=1 # is there a bug in Fuseki?!
+BSBM_CONCURRENT_CLIENTS=4
+BSBM_SEED=1212123
 BSBM_QUERY_TIMEOUT=30000
 
 TDB_LOADER=tdbloader
@@ -102,7 +103,7 @@ fi
 run_fuseki() {
     echo "Running Fuseki..."
     kill `ps -ef | grep fuseki | grep -v grep | awk '{print $2}'`
-    java -jar $BSBM_ROOT_PATH/fuseki/target/fuseki-0.2.0-SNAPSHOT-sys.jar --update --loc=$BSBM_ROOT_PATH/datasets/tdb-$BSBM_SCALE_FACTOR/TDB /bsbm &> /dev/null &
+    java -jar $BSBM_ROOT_PATH/fuseki/target/fuseki-0.2.0-SNAPSHOT-sys.jar --update --loc=$BSBM_ROOT_PATH/datasets/tdb-$BSBM_SCALE_FACTOR/TDB /bsbm &>> $BSBM_ROOT_PATH/results/fuseki.log &
     sleep 4
     echo "done."
 }
@@ -129,6 +130,19 @@ if [ ! -f "$BSBM_ROOT_PATH/results/bsbm-results-$BSBM_SCALE_FACTOR-fuseki-update
     fi
     cd $BSBM_ROOT_PATH/bsbmtools
     java -cp "lib/*" -Xmx256M benchmark.testdriver.TestDriver -runs $BSBM_NUM_QUERY_MIXES -w $BSBM_NUM_QUERY_WARM_UP -mt $BSBM_CONCURRENT_CLIENTS -t $BSBM_QUERY_TIMEOUT -ucf usecases/explore/sparql.txt -u http://127.0.0.1:3030/bsbm/update -udataset $BSBM_ROOT_PATH/datasets/bsbm-dataset-$BSBM_SCALE_FACTOR/dataset_update.nt http://127.0.0.1:3030/bsbm/query > $BSBM_ROOT_PATH/results/bsbm-results-$BSBM_SCALE_FACTOR-fuseki-update.txt
+    kill `ps -ef | grep fuseki | grep -v grep | awk '{print $2}'`
+    echo "done."
+fi
+
+
+if [ ! -f "$BSBM_ROOT_PATH/results/bsbm-results-$BSBM_SCALE_FACTOR-fuseki-bi.txt" ]; then
+    run_fuseki
+    echo "Running BSBM benchmark (scale factor is $BSBM_SCALE_FACTOR) with BI against Fuseki..."
+    if [ ! -d "$BSBM_ROOT_PATH/results" ]; then
+        mkdir $BSBM_ROOT_PATH/results
+    fi
+    cd $BSBM_ROOT_PATH/bsbmtools
+    java -cp "lib/*" -Xmx256M benchmark.testdriver.TestDriver -runs $BSBM_NUM_QUERY_MIXES -w $BSBM_NUM_QUERY_WARM_UP -mt $BSBM_CONCURRENT_CLIENTS -t $BSBM_QUERY_TIMEOUT -ucf usecases/businessIntelligence/sparql.txt -u http://127.0.0.1:3030/bsbm/update -udataset $BSBM_ROOT_PATH/datasets/bsbm-dataset-$BSBM_SCALE_FACTOR/dataset_update.nt http://127.0.0.1:3030/bsbm/query > $BSBM_ROOT_PATH/results/bsbm-results-$BSBM_SCALE_FACTOR-fuseki-bi.txt
     kill `ps -ef | grep fuseki | grep -v grep | awk '{print $2}'`
     echo "done."
 fi
